Updated Project with Serialiazble added. Changes:
added a toString() method in both the Conference and Paper class.
Sincethese are new object created outside of the java library, there is
no natural toString for it and if you display what you deserialize, it
will only print the memory location.

Stevan I turned your mock data inside of the GUImain into a serialized
form so I could test it.

further actions:

Stevan- i am not seeing the action listeners for adding roles and
papers, just for exiting  and changing the role.  Ideally, I feel the
best way to go about this is to serialize every Paper and Conference
that we create.  The only issue (minor) I can foresee in our
serialization process is that different components of the conference are
being updated at different times, mainly the user.  So we need not only
serialize our conference everytime a user is being created (i.e. a new
conference is created or a new user is created), but also every time
information is being updated (i.e. a user adds on a new role, etc).  So
we need to serialize on either every action button, or to make things
simplier, we could also serialize on every application exit.  Now that I
am thinking about it, it might be best to serialize on every application
exit and deserialize on every application open.